+++ ChangeLog.md
@@ -11,17 +11,38 @@ A package of FFVC with related libraries.
 |Intel / Intel 17.0.1 ||ok^(1)|Apple MacBook Pro OSX 10.12.4|
 |Intel / GNU 6.2.0    ||ok^(1)|Apple MacBook Pro OSX 10.12.4|
 |Fujitsu / fx10       ||ok|Hayaka /opt/FJSVfxlang/1.2.1|
-|Fujitsu / fx100      |||
-|Fujitsu / K          |||
+|Fujitsu / fx100      ||ok|
+|Fujitsu / K          ||ok|
 
  - ^(1) OpenMPI 1.10.4
 
 
 ## RELEASE HISTORY
+---
+- 2017-07-06 Version 2.5.2
+  - modify install.sh
+  - FFVC-2.5.2
+  - TextParser-1.8.5
+  - PMlib-5.6.6
+  - Polylib-3.6.7
+  - CPMlib-2.4.4
+  - CDMlib-1.1.3
+  
+
+---
+- 2017-07-05 Version 2.5.1b
+  - modify install.sh
+  - TextParser-1.8.4
+  - PMlib-5.6.5
+  - Polylib-3.6.5
+  - CPMlib-2.4.3
+  - CDMlib-1.1.2
+
   
 ---
 - 2017-04-24 Version 2.5.1
   - FFVC-2.5.1
+
 
 ---
 - 2017-04-22 Version 2.5.0